# swagger.json
{"openapi":"3.0.0","paths":{"/v1/create/actor":{"post":{"operationId":"FilmController_executeCreateActor","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateActorDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"success":true,"message":{"id":"ecf0a010-482b-4bf0-8ba3-e3a002fdfb7a"}}}}}}},"tags":["v1/create/actor"]}},"/v1/get/actor":{"get":{"operationId":"FilmController_getActor","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateActorDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"success":true,"message":{"nombre":"Luke Skywalker","estatura":"172","peso":"77","colorCabello":"blond","colorPiel":"fair","colorOjos":"blue","fechaCumpleanhos":"19BBY","genero":"male","mundoCasa":"https://swapi.py4e.com/api/planets/1/","peliculas":["https://swapi.py4e.com/api/films/1/"],"especies":[{"nombre":"Human","clasificacion":"mammal","designacion":"sentient","estatura_promedio":"180","colores_piel":"caucasian, black, asian, hispanic","colores_cabello":"blonde, brown, black, red","colores_ojos":"brown, blue, green, hazel, grey, amber","esperanza_vida":"120","lenguaje":"Galactic Basic","personas":[],"fechaCreacion":"2014-12-10T13:52:11.567000Z","fechaModificacion":"2014-12-20T21:36:42.136000Z","url":"https://swapi.py4e.com/api/species/1/"}],"vehiculos":[{"nombre":"Snowspeeder","modelo":"t-47 airspeeder","fabricante":"Incom corporation","costoEnCreditos":"unknown","longitud":"4.5","maxima_velocidad_admosferica":"650","multitud":"2","pasajeros":"0","capacidad_carga":"10","consumibles":"none","clase_vehiculo":"airspeeder","piloto":["https://swapi.py4e.com/api/people/1/","https://swapi.py4e.com/api/people/18/"]}],"navesEstelares":["https://swapi.py4e.com/api/starships/12/","https://swapi.py4e.com/api/starships/22/"],"fechaCreacion":"2014-12-09T13:50:51.644000Z","fechaModificacion":"2014-12-20T21:17:56.891000Z","url":"https://swapi.py4e.com/api/people/1/"}}}}}}},"tags":["v1/create/actor"]}}},"info":{"title":"Film example","description":"The Film API description","version":"1.0","contact":{}},"tags":[{"name":"film","description":""}],"servers":[],"components":{"schemas":{"CreateActorDto":{"type":"object","properties":{"nombre":{"type":"string","description":"nombre de personaje"},"estatura":{"type":"string"},"peso":{"type":"string"},"colorCabello":{"type":"string"},"colorPiel":{"type":"string"},"colorOjos":{"type":"string"},"fechaCumpleanhos":{"type":"string"},"genero":{"type":"string"}},"required":["nombre","estatura","peso","colorCabello","colorPiel","colorOjos","fechaCumpleanhos","genero"]}}}}